ingestion-api/pyproject.toml
[project]
name = "ingestion-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.14",
    "greenlet>=3.2.3",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "openai>=1.92.3",
    "pgvector>=0.4.1",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pypdf>=5.6.1",
    "pypdf2>=3.0.1",
    "sqlalchemy>=2.0.41",
    "structlog>=24.1.0",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
]
